Enhance channel

+++ app/controllers/projects/attachments_controller.rb
@@ -27,8 +27,9 @@ class Projects::AttachmentsController < Projects::ApplicationController
   end
 
   def destroy
-    event_service.delete_attachment(@attachment, current_user)
-    @attachment.destroy
+    if @attachment.destroy
+      event_service.delete_attachment(@attachment, current_user)
+    end
     respond_with @attachment, location: project_attachments_path
   end
